Optimize RGS hot path with zero-alloc PRNG, fixed-point money, and wallet breakers.
Use int64 minor units end-to-end for bets and payouts, sync.Pool for spin state, and per-merchant go-zero breakers that trip on >500ms wallet latency.

+++ internal/model/gameRoundReplayModel.go
@@ -17,7 +17,7 @@ type GameRoundReplay struct {
 	ServerSeed   string    `db:"server_seed"`
 	ClientSeed   string    `db:"client_seed"`
 	Nonce        string    `db:"nonce"`
-	BetAmount    float64   `db:"bet_amount"`
+	BetAmount    int64     `db:"bet_amount"`
 	SequenceID   uint64    `db:"sequence_id"`
 	CreatedAt    time.Time `db:"created_at"`
 }
